clean up variable defs, add test
