--- funnel/collection/config.py
@@ -1,1 +1,0 @@
-derived_variable_registry = dict()
