+++ funnel/collection/registry.py
@@ -1,0 +1,2 @@
+derived_variable_registry = {}
+query_dependent_operator_registry = {}
